Corrects Svelte component watch/reload
When using Vite development server, Svelte's preference for `lib`
directory was incorrectly ignored, causing anything with `/lib/` in the
path to be excluded.

This was in place due Shards usage of `lib` to install dependencies and
the endless nested symlinks.

Adjusted the ignore rules to exclude things only relative to the root of
the server.